requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart